Changed time data from double to int
Timing methods were changed to use System.nanoTime (in a previous commit), which returns long. The times are relatively small, so the time data was changed to int.

+++ src/MyFiles.java
@@ -92,7 +92,7 @@ class MyFiles {
 		}
 	}
 	
-	public static void compileTestData(double[][] data, int[] seqCutoffs, String dataSize, String seqPar) {
+	public static void compileTestData(int[][] data, int[] seqCutoffs, String dataSize, String seqPar) {
 //		String path = "";
 		String filename = seqPar + "_" + dataSize + "-data" + ".txt";
 		try {
